[Refactor] gcTime, staleTime 을 자정까지 남은 시간으로 설정 및 date 유틸 파일 생성
이유: 쇼츠는 자정마다 업데이트될 것이기 때문
